pyproject.toml
[project]
name = "langchain-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=1.0.4",
    "langchain-openai>=1.0.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
